Change returned type of get_axis method

+++ src/ngraph/op/gather.hpp
@@ -70,7 +70,7 @@ namespace ngraph
                        const Output<Node>& axis);
 
                 size_t get_version() const override { return 1; }
-                int64_t get_axis() const;
+                size_t get_axis() const;
 
                 void validate_and_infer_types() override;
